+++ src/GitHub.hs
@@ -16,9 +16,16 @@ module GitHub (
     -- | See <https://developer.github.com/v3/activity/>
 
     -- ** Events
-    -- | See https://developer.github.com/v3/activity/events/#events
+    -- | See <https://developer.github.com/v3/activity/events/>
     repositoryEventsR,
     userEventsR,
+
+    -- ** Notifications
+    -- | See <https://developer.github.com/v3/activity/notifications/>
+    getNotificationsR,
+    markNotificationAsReadR,
+    markAllNotificationsAsReadR,
+
     -- ** Starring
     -- | See <https://developer.github.com/v3/activity/starring/>
     --
@@ -374,6 +381,7 @@ module GitHub (
 
 import GitHub.Data
 import GitHub.Endpoints.Activity.Events
+import GitHub.Endpoints.Activity.Notifications
 import GitHub.Endpoints.Activity.Starring
 import GitHub.Endpoints.Activity.Watching
 import GitHub.Endpoints.Gists