+++ src/GitHub/Endpoints/Activity/Notifications.hs
@@ -15,25 +15,31 @@ import Prelude ()
 
 getNotifications :: Auth -> IO (Either Error (Vector Notification))
 getNotifications auth =
-  executeRequest auth $ getNotificationsR FetchAll
+    executeRequest auth $ getNotificationsR FetchAll
 
+-- | List your notifications.
+-- See <https://developer.github.com/v3/activity/notifications/#list-your-notifications>
 getNotificationsR :: FetchCount -> Request 'RA (Vector Notification)
-getNotificationsR =
-  pagedQuery ["notifications"] []
+getNotificationsR = pagedQuery ["notifications"] []
 
 markNotificationAsRead :: Auth -> Id Notification -> IO (Either Error ())
-markNotificationAsRead auth notificationId =
-  executeRequest auth $ markNotificationAsReadR notificationId
+markNotificationAsRead auth nid =
+    executeRequest auth $ markNotificationAsReadR nid
 
-markNotificationAsReadR :: Id Notification -> Request 'RW ()
-markNotificationAsReadR notificationId = SimpleQuery $
-  Command Patch ["notifications", "threads", toPathPart notificationId]
-            (encode ())
+-- | Mark a thread as read.
+-- See <https://developer.github.com/v3/activity/notifications/#mark-a-thread-as-read>
+markNotificationAsReadR :: Id Notification -> GenRequest 'MtUnit 'RW ()
+markNotificationAsReadR nid = Command
+    Patch
+    ["notifications", "threads", toPathPart nid]
+    (encode ())
 
 markNotificationsAsRead :: Auth -> IO (Either Error ())
 markNotificationsAsRead auth =
   executeRequest auth markAllNotificationsAsReadR
 
-markAllNotificationsAsReadR :: Request 'RW ()
-markAllNotificationsAsReadR = SimpleQuery $
-  Command Put ["notifications"] $ encode emptyObject
+-- | Mark as read.
+-- See <https://developer.github.com/v3/activity/notifications/#mark-as-read>
+markAllNotificationsAsReadR :: GenRequest 'MtUnit 'RW ()
+markAllNotificationsAsReadR =
+    Command Put ["notifications"] $ encode emptyObject